+++ plugins/faustwp/tests/acceptance/PostPreviewCest.php
@@ -28,7 +28,7 @@ class PostPreviewCest
         $I->wait(4); // Wait for previewlinks.js to modify button href.
         $I->seeLink(
             'Preview in new tab',
-            "${front_end_url}/${post_name}/?preview=true&p=${post_id}",
+            "${front_end_url}/${post_name}/?preview=true&p=${post_id}&typeName=Post",
         );
 
 		$I->click('Preview in new tab');